service/requirements.txt
Flask
flask_cors
numpy
torch
torchvision
pillow